Remove the saddle

+++ app/models/StatRow.scala
@@ -1,5 +1,5 @@
 package models
 
-case class StatRow(team:Team, rank:Int, value:Double, percentile:Double, zScore:Double) {
+case class StatRow(team:Team, rank:Double, value:Double, percentile:Double, zScore:Double) {
 
 }
